Fix: Read CORS_ALLOWED_ORIGINS from environment variable
- SecurityConfig now uses app.cors.allowed-origins property
- Parses comma-separated origins from CORS_ALLOWED_ORIGINS env var
- Also fixes tps column type (DECIMAL to DOUBLE PRECISION)

+++ scripts/db/schema/001_initial_schema.sql
@@ -75,7 +75,7 @@ CREATE TABLE IF NOT EXISTS solves (
     tags VARCHAR(255)[],
     -- Performance tracking fields
     move_count INTEGER,
-    tps DECIMAL(5,2), -- Turns per second
+    tps DOUBLE PRECISION, -- Turns per second
     -- Future analytics fields
     cross_time_ms INTEGER,
     f2l_time_ms INTEGER,
